# openapi/clhear-v1.yaml
components:
  schemas:
...
    ContributionProposal:
      additionalProperties: false
      description: 'Public parsing rules and clause mappings.


        List source keys, adapter definitions, normalizer rules, and structural

        elements (an element id, a clause reference, and a provenance hash).

        Organisation facts, live controls, owners, evidence locators, database

        URLs, and credentials are rejected. Validating this document sends it nowhere.'
      properties:
        adapter_definitions:
          items:
            $ref: '#/components/schemas/AdapterDefinition'
          title: Adapter Definitions
          type: array
        normalizer_rules:
          items:
            $ref: '#/components/schemas/NormalizerRule'
          title: Normalizer Rules
          type: array
        source_keys:
          items:
            type: string
          title: Source Keys
          type: array
        structural_elements:
          items:
            $ref: '#/components/schemas/StructuralElement'
          title: Structural Elements
          type: array
      required:
      - source_keys
      title: ContributionProposal
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ProfileBody:
      additionalProperties: false
      properties:
        attributes:
          additionalProperties: true
          title: Attributes
          type: object
        name:
          default: ''
          title: Name
          type: string
      required:
      - attributes
      title: ProfileBody
      type: object
...
info:
  description: One CLHEAR install. Register the texts you choose, describe the organisation,
    and POST /v1/runs. The call returns a run id. The worker writes the blueprint.
    Read items, explanations, and clause pointers from the release. ContributionProposal
    is checked locally and is sent nowhere.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: CLHEAR
  version: 0.1.0
openapi: 3.1.0
paths:
  /v1/adapters:
    get:
      operationId: adapters_v1_adapters_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Adapters V1 Adapters Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Adapters
  /v1/blueprints/{blueprint_id}:
    get:
      operationId: get_blueprint_v1_blueprints__blueprint_id__get
      parameters:
      - in: path
        name: blueprint_id
        required: true
        schema:
          title: Blueprint Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Blueprint V1 Blueprints  Blueprint Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Blueprint
  /v1/blueprints/{blueprint_id}/diff:
    get:
      operationId: diff_blueprint_v1_blueprints__blueprint_id__diff_get
      parameters:
      - in: path
        name: blueprint_id
        required: true
        schema:
          title: Blueprint Id
          type: string
      - in: query
        name: against
        required: true
        schema:
          title: Against
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Diff Blueprint V1 Blueprints  Blueprint Id  Diff Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Diff Blueprint
  /v1/contributions/validate:
    post:
      description: Validate a contribution proposal. This service does not send it
        anywhere.
      operationId: validate_contribution_v1_contributions_validate_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionProposal'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionProposal'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Validate Contribution
...
  /v1/profile-schema:
    get:
      operationId: profile_schema_v1_profile_schema_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Profile Schema V1 Profile Schema Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Profile Schema
  /v1/profiles/{profile_id}:
    get:
      operationId: get_profile_v1_profiles__profile_id__get
      parameters:
      - in: path
        name: profile_id
        required: true
        schema:
          title: Profile Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Profile V1 Profiles  Profile Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Profile
    put:
      operationId: put_profile_v1_profiles__profile_id__put
      parameters:
      - in: path
        name: profile_id
        required: true
        schema:
          title: Profile Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Put Profile V1 Profiles  Profile Id  Put
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Put Profile
  /v1/releases/{release_id}:
    get:
      operationId: get_release_v1_releases__release_id__get
      parameters:
      - in: path
        name: release_id
        required: true
        schema:
          title: Release Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Release V1 Releases  Release Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Release
  /v1/releases/{release_id}/blueprints/{profile_id}:
    get:
      operationId: get_release_blueprint_v1_releases__release_id__blueprints__profile_id__get
      parameters:
      - in: path
        name: release_id
        required: true
        schema:
          title: Release Id
          type: string
      - in: path
        name: profile_id
        required: true
        schema:
          title: Profile Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Release Blueprint V1 Releases  Release Id  Blueprints  Profile
                  Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Release Blueprint
  /v1/runs:
    post:
      operationId: create_run_v1_runs_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunBody'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Run V1 Runs Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Run
  /v1/runs/{run_id}:
    get:
      operationId: get_run_v1_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Run V1 Runs  Run Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Run
  /v1/runs/{run_id}/logs:
    get:
      operationId: get_logs_v1_runs__run_id__logs_get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Logs V1 Runs  Run Id  Logs Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Logs
  /v1/scopes:
    get:
      operationId: list_scopes_v1_scopes_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response List Scopes V1 Scopes Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Scopes
    post:
      operationId: create_scope_v1_scopes_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScopeBody'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Scope V1 Scopes Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Scope
  /v1/scopes/{name}:
    get:
      operationId: get_scope_v1_scopes__name__get
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Scope V1 Scopes  Name  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Scope
  /v1/sources:
    get:
      operationId: list_sources_v1_sources_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response List Sources V1 Sources Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Sources
    post:
      operationId: create_source_v1_sources_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Source V1 Sources Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Source
...
  /v1/sources/{key}/test-fetch:
    post:
      operationId: test_fetch_v1_sources__key__test_fetch_post
      parameters:
      - in: path
        name: key
        required: true
        schema:
          title: Key
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Test Fetch V1 Sources  Key  Test Fetch Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Test Fetch
...